Move frauds tests into a separate file
