+++ solidity/test/bridge/Bridge.test.ts
@@ -4,12 +4,7 @@
 import { ethers, helpers, waffle } from "hardhat"
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers"
 import chai, { expect } from "chai"
-import {
-  BigNumber,
-  BigNumberish,
-  ContractTransaction,
-  Transaction,
-} from "ethers"
+import { BigNumber, BigNumberish, ContractTransaction } from "ethers"
 import { BytesLike } from "@ethersproject/bytes"
 import { smock } from "@defi-wonderland/smock"
 import type { FakeContract } from "@defi-wonderland/smock"
@@ -55,16 +50,6 @@ import {
   MultiplePendingRequestedRedemptionsWithProvablyUnspendable,
   MultiplePendingRequestedRedemptionsWithMultipleInputs,
 } from "../data/redemption"
-import {
-  fraudWalletPublicKey,
-  fraudWalletPublicKeyHash,
-  nonWitnessSignSingleInputTx,
-  nonWitnessSignMultipleInputsTx,
-  witnessSignSingleInputTx,
-  witnessSignMultipleInputTx,
-  wrongSighashType,
-} from "../data/fraud"
-import { ecdsaWalletTestData } from "../data/ecdsa"
 import { walletState } from "../fixtures"
 import { Wallets__factory } from "../../typechain"
 
@@ -4860,1388 +4845,6 @@ describe("Bridge", () => {
     })
   })
 
-  describe("submitFraudChallenge", () => {
-    const data = witnessSignSingleInputTx
-
-    context("when the wallet is in Live state", () => {
-      context("when the amount of ETH deposited is enough", () => {
-        context(
-          "when the data needed for signature verification is correct",
-          () => {
-            context("when the fraud challenge does not exist yet", () => {
-              let tx: Transaction
-
-              before(async () => {
-                await createSnapshot()
-
-                await bridge.setWallet(fraudWalletPublicKeyHash, {
-                  ecdsaWalletID: ethers.constants.HashZero,
-                  mainUtxoHash: ethers.constants.HashZero,
-                  pendingRedemptionsValue: 0,
-                  createdAt: await lastBlockTime(),
-                  state: walletState.Live,
-                })
-
-                tx = await bridge
-                  .connect(thirdParty)
-                  .submitFraudChallenge(
-                    fraudWalletPublicKey,
-                    data.sighash,
-                    data.signature,
-                    {
-                      value: await bridge.fraudChallengeDepositAmount(),
-                    }
-                  )
-              })
-
-              after(async () => {
-                await restoreSnapshot()
-              })
-
-              it("should transfer ether from the caller to the bridge", async () => {
-                await expect(tx).to.changeEtherBalance(
-                  thirdParty,
-                  (await bridge.fraudChallengeDepositAmount()).mul(-1)
-                )
-                await expect(tx).to.changeEtherBalance(
-                  bridge,
-                  await bridge.fraudChallengeDepositAmount()
-                )
-              })
-
-              it("should store the fraud challenge data", async () => {
-                const challengeKey = buildChallengeKey(
-                  fraudWalletPublicKey,
-                  data.sighash,
-                  data.signature.v,
-                  data.signature.r,
-                  data.signature.s
-                )
-
-                const fraudChallenge = await bridge.fraudChallenges(
-                  challengeKey
-                )
-
-                expect(fraudChallenge.challenger).to.equal(
-                  await thirdParty.getAddress()
-                )
-                expect(fraudChallenge.depositAmount).to.equal(
-                  await bridge.fraudChallengeDepositAmount()
-                )
-                expect(fraudChallenge.reportedAt).to.equal(
-                  await lastBlockTime()
-                )
-                expect(fraudChallenge.closed).to.equal(false)
-              })
-
-              it("should emit FraudChallengeSubmitted event", async () => {
-                await expect(tx)
-                  .to.emit(bridge, "FraudChallengeSubmitted")
-                  .withArgs(
-                    fraudWalletPublicKeyHash,
-                    data.sighash,
-                    data.signature.v,
-                    data.signature.r,
-                    data.signature.s
-                  )
-              })
-            })
-
-            context("when the fraud challenge already exists", () => {
-              before(async () => {
-                await createSnapshot()
-
-                await bridge.setWallet(fraudWalletPublicKeyHash, {
-                  ecdsaWalletID: ethers.constants.HashZero,
-                  mainUtxoHash: ethers.constants.HashZero,
-                  pendingRedemptionsValue: 0,
-                  createdAt: await lastBlockTime(),
-                  state: walletState.Live,
-                })
-
-                await bridge
-                  .connect(thirdParty)
-                  .submitFraudChallenge(
-                    fraudWalletPublicKey,
-                    data.sighash,
-                    data.signature,
-                    {
-                      value: await bridge.fraudChallengeDepositAmount(),
-                    }
-                  )
-              })
-
-              after(async () => {
-                await restoreSnapshot()
-              })
-
-              it("should revert", async () => {
-                await expect(
-                  bridge
-                    .connect(thirdParty)
-                    .submitFraudChallenge(
-                      fraudWalletPublicKey,
-                      data.sighash,
-                      data.signature,
-                      {
-                        value: await bridge.fraudChallengeDepositAmount(),
-                      }
-                    )
-                ).to.be.revertedWith("Fraud challenge already exists")
-              })
-            })
-          }
-        )
-
-        context("when incorrect wallet public key is used", () => {
-          const incorrectWalletPublicKey =
-            "0xffc045ade19f8a5d464299146ce069049cdcc2390a9b44d9abcd83f11d8cce4" +
-            "01ea6800e307b87aadebdcd2f7293cc60f0526afaff1a7b1abddfd787e6c5871e"
-
-          const incorrectWalletPublicKeyHash =
-            "0xb5222794425b9b8cd8c3358e73a50dea73480927"
-
-          before(async () => {
-            await createSnapshot()
-            await bridge.setWallet(incorrectWalletPublicKeyHash, {
-              ecdsaWalletID: ethers.constants.HashZero,
-              mainUtxoHash: ethers.constants.HashZero,
-              pendingRedemptionsValue: 0,
-              createdAt: await lastBlockTime(),
-              state: walletState.Live,
-            })
-          })
-
-          after(async () => {
-            await restoreSnapshot()
-          })
-
-          it("should revert", async () => {
-            await expect(
-              bridge
-                .connect(thirdParty)
-                .submitFraudChallenge(
-                  incorrectWalletPublicKey,
-                  data.sighash,
-                  data.signature,
-                  {
-                    value: await bridge.fraudChallengeDepositAmount(),
-                  }
-                )
-            ).to.be.revertedWith("Signature verification failure")
-          })
-        })
-
-        context("when incorrect sighash is used", () => {
-          const incorrectSighash =
-            "0x9e8e249791a5636e5e007fc15487b5a5bd6e60f73f7e236a7025cd63b904650b"
-
-          before(async () => {
-            await createSnapshot()
-            await bridge.setWallet(fraudWalletPublicKeyHash, {
-              ecdsaWalletID: ethers.constants.HashZero,
-              mainUtxoHash: ethers.constants.HashZero,
-              pendingRedemptionsValue: 0,
-              createdAt: await lastBlockTime(),
-              state: walletState.Live,
-            })
-          })
-
-          after(async () => {
-            await restoreSnapshot()
-          })
-
-          it("should revert", async () => {
-            await expect(
-              bridge
-                .connect(thirdParty)
-                .submitFraudChallenge(
-                  fraudWalletPublicKey,
-                  incorrectSighash,
-                  data.signature,
-                  {
-                    value: await bridge.fraudChallengeDepositAmount(),
-                  }
-                )
-            ).to.be.revertedWith("Signature verification failure")
-          })
-        })
-
-        context("when incorrect recovery ID is used", () => {
-          const incorrectV = data.signature.v + 1
-
-          before(async () => {
-            await createSnapshot()
-            await bridge.setWallet(fraudWalletPublicKeyHash, {
-              ecdsaWalletID: ethers.constants.HashZero,
-              mainUtxoHash: ethers.constants.HashZero,
-              pendingRedemptionsValue: 0,
-              createdAt: await lastBlockTime(),
-              state: walletState.Live,
-            })
-          })
-
-          after(async () => {
-            await restoreSnapshot()
-          })
-
-          it("should revert", async () => {
-            await expect(
-              bridge.connect(thirdParty).submitFraudChallenge(
-                fraudWalletPublicKey,
-                data.sighash,
-                {
-                  r: data.signature.r,
-                  s: data.signature.s,
-                  v: incorrectV,
-                },
-                {
-                  value: await bridge.fraudChallengeDepositAmount(),
-                }
-              )
-            ).to.be.revertedWith("Signature verification failure")
-          })
-        })
-
-        context("when incorrect signature data is used", () => {
-          // just swap r and s
-          const incorrectS = data.signature.r
-          const incorrectR = data.signature.s
-
-          before(async () => {
-            await createSnapshot()
-            await bridge.setWallet(fraudWalletPublicKeyHash, {
-              ecdsaWalletID: ethers.constants.HashZero,
-              mainUtxoHash: ethers.constants.HashZero,
-              pendingRedemptionsValue: 0,
-              createdAt: await lastBlockTime(),
-              state: walletState.Live,
-            })
-          })
-
-          after(async () => {
-            await restoreSnapshot()
-          })
-
-          it("should revert", async () => {
-            await expect(
-              bridge.connect(thirdParty).submitFraudChallenge(
-                fraudWalletPublicKey,
-                data.sighash,
-                {
-                  r: incorrectR,
-                  s: incorrectS,
-                  v: data.signature.v,
-                },
-                {
-                  value: await bridge.fraudChallengeDepositAmount(),
-                }
-              )
-            ).to.be.revertedWith("Signature verification failure")
-          })
-        })
-      })
-
-      context("when the amount of ETH deposited is too low", () => {
-        before(async () => {
-          await createSnapshot()
-          await bridge.setWallet(fraudWalletPublicKeyHash, {
-            ecdsaWalletID: ethers.constants.HashZero,
-            mainUtxoHash: ethers.constants.HashZero,
-            pendingRedemptionsValue: 0,
-            createdAt: await lastBlockTime(),
-            state: walletState.Live,
-          })
-        })
-
-        after(async () => {
-          await restoreSnapshot()
-        })
-
-        it("should revert", async () => {
-          await expect(
-            bridge
-              .connect(thirdParty)
-              .submitFraudChallenge(
-                fraudWalletPublicKey,
-                data.sighash,
-                data.signature,
-                {
-                  value: (await bridge.fraudChallengeDepositAmount()).sub(1),
-                }
-              )
-          ).to.be.revertedWith("The amount of ETH deposited is too low")
-        })
-      })
-    })
-
-    context("when the wallet is in MovingFunds state", () => {
-      before(async () => {
-        await createSnapshot()
-        await bridge.setWallet(fraudWalletPublicKeyHash, {
-          ecdsaWalletID: ethers.constants.HashZero,
-          mainUtxoHash: ethers.constants.HashZero,
-          pendingRedemptionsValue: 0,
-          createdAt: await lastBlockTime(),
-          state: walletState.Live,
-        })
-      })
-
-      after(async () => {
-        await restoreSnapshot()
-      })
-
-      it("should succeed", async () => {
-        await expect(
-          bridge
-            .connect(thirdParty)
-            .submitFraudChallenge(
-              fraudWalletPublicKey,
-              data.sighash,
-              data.signature,
-              {
-                value: await bridge.fraudChallengeDepositAmount(),
-              }
-            )
-        ).to.not.be.reverted
-      })
-    })
-
-    context("when the wallet is in neither Live nor MovingFunds state", () => {
-      before(async () => {
-        await createSnapshot()
-        await bridge.setWallet(fraudWalletPublicKeyHash, {
-          ecdsaWalletID: ethers.constants.HashZero,
-          mainUtxoHash: ethers.constants.HashZero,
-          pendingRedemptionsValue: 0,
-          createdAt: await lastBlockTime(),
-          state: walletState.Closed,
-        })
-      })
-
-      after(async () => {
-        await restoreSnapshot()
-      })
-
-      it("should revert", async () => {
-        await expect(
-          bridge
-            .connect(thirdParty)
-            .submitFraudChallenge(
-              fraudWalletPublicKey,
-              data.sighash,
-              data.signature,
-              {
-                value: await bridge.fraudChallengeDepositAmount(),
-              }
-            )
-        ).to.be.revertedWith("Wallet is neither in Live nor MovingFunds state")
-      })
-    })
-  })
-
-  describe("defeatFraudChallenge", () => {
-    context("when the challenge exists", () => {
-      context("when the challenge is open", () => {
-        context("when the sighash type is correct", () => {
-          context("when the input is non-witness", () => {
-            context("when the transaction has single input", () => {
-              context(
-                "when the inputs are marked correctly spent in the Bridge",
-                () => {
-                  const data = nonWitnessSignSingleInputTx
-                  let tx: Transaction
-
-                  before(async () => {
-                    await createSnapshot()
-
-                    await bridge.setWallet(fraudWalletPublicKeyHash, {
-                      ecdsaWalletID: ethers.constants.HashZero,
-                      mainUtxoHash: ethers.constants.HashZero,
-                      pendingRedemptionsValue: 0,
-                      createdAt: await lastBlockTime(),
-                      state: walletState.Live,
-                    })
-                    await bridge.setSweptDeposits(data.deposits)
-                    await bridge.setSpentMainUtxos(data.spentMainUtxos)
-
-                    await bridge
-                      .connect(thirdParty)
-                      .submitFraudChallenge(
-                        fraudWalletPublicKey,
-                        data.sighash,
-                        data.signature,
-                        {
-                          value: await bridge.fraudChallengeDepositAmount(),
-                        }
-                      )
-
-                    tx = await bridge
-                      .connect(thirdParty)
-                      .defeatFraudChallenge(
-                        fraudWalletPublicKey,
-                        data.preimage,
-                        data.signature,
-                        false
-                      )
-                  })
-
-                  after(async () => {
-                    await restoreSnapshot()
-                  })
-
-                  it("should mark the challenge as closed", async () => {
-                    const challengeKey = buildChallengeKey(
-                      fraudWalletPublicKey,
-                      data.sighash,
-                      data.signature.v,
-                      data.signature.r,
-                      data.signature.s
-                    )
-
-                    const fraudChallenge = await bridge.fraudChallenges(
-                      challengeKey
-                    )
-
-                    expect(fraudChallenge.closed).to.equal(true)
-                  })
-
-                  it("should send the ether deposited by the challenger to the treasury", async () => {
-                    await expect(tx).to.changeEtherBalance(
-                      bridge,
-                      (await bridge.fraudChallengeDepositAmount()).mul(-1)
-                    )
-                    await expect(tx).to.changeEtherBalance(
-                      treasury,
-                      await bridge.fraudChallengeDepositAmount()
-                    )
-                  })
-
-                  it("should emit FraudChallengeDefeated event", async () => {
-                    await expect(tx)
-                      .to.emit(bridge, "FraudChallengeDefeated")
-                      .withArgs(
-                        fraudWalletPublicKeyHash,
-                        data.sighash,
-                        data.signature.v,
-                        data.signature.r,
-                        data.signature.s
-                      )
-                  })
-                }
-              )
-
-              context(
-                "when the inputs are not marked as correctly spent in the Bridge",
-                () => {
-                  const data = nonWitnessSignSingleInputTx
-
-                  before(async () => {
-                    await createSnapshot()
-
-                    await bridge.setWallet(fraudWalletPublicKeyHash, {
-                      ecdsaWalletID: ethers.constants.HashZero,
-                      mainUtxoHash: ethers.constants.HashZero,
-                      pendingRedemptionsValue: 0,
-                      createdAt: await lastBlockTime(),
-                      state: walletState.Live,
-                    })
-
-                    await bridge
-                      .connect(thirdParty)
-                      .submitFraudChallenge(
-                        fraudWalletPublicKey,
-                        data.sighash,
-                        data.signature,
-                        {
-                          value: await bridge.fraudChallengeDepositAmount(),
-                        }
-                      )
-                  })
-
-                  after(async () => {
-                    await restoreSnapshot()
-                  })
-
-                  it("should revert", async () => {
-                    await expect(
-                      bridge
-                        .connect(thirdParty)
-                        .defeatFraudChallenge(
-                          fraudWalletPublicKey,
-                          data.preimage,
-                          data.signature,
-                          false
-                        )
-                    ).to.be.revertedWith(
-                      "Spent UTXO not found among correctly spent UTXOs"
-                    )
-                  })
-                }
-              )
-            })
-
-            context("when the transaction has multiple inputs", () => {
-              context(
-                "when the inputs are marked correctly spent in the Bridge",
-                () => {
-                  const data = nonWitnessSignMultipleInputsTx
-                  let tx: Transaction
-
-                  before(async () => {
-                    await createSnapshot()
-
-                    await bridge.setWallet(fraudWalletPublicKeyHash, {
-                      ecdsaWalletID: ethers.constants.HashZero,
-                      mainUtxoHash: ethers.constants.HashZero,
-                      pendingRedemptionsValue: 0,
-                      createdAt: await lastBlockTime(),
-                      state: walletState.Live,
-                    })
-                    await bridge.setSweptDeposits(data.deposits)
-                    await bridge.setSpentMainUtxos(data.spentMainUtxos)
-
-                    await bridge
-                      .connect(thirdParty)
-                      .submitFraudChallenge(
-                        fraudWalletPublicKey,
-                        data.sighash,
-                        data.signature,
-                        {
-                          value: await bridge.fraudChallengeDepositAmount(),
-                        }
-                      )
-
-                    tx = await bridge
-                      .connect(thirdParty)
-                      .defeatFraudChallenge(
-                        fraudWalletPublicKey,
-                        data.preimage,
-                        data.signature,
-                        false
-                      )
-                  })
-
-                  after(async () => {
-                    await restoreSnapshot()
-                  })
-
-                  it("should mark the challenge as closed", async () => {
-                    const challengeKey = buildChallengeKey(
-                      fraudWalletPublicKey,
-                      data.sighash,
-                      data.signature.v,
-                      data.signature.r,
-                      data.signature.s
-                    )
-
-                    const fraudChallenge = await bridge.fraudChallenges(
-                      challengeKey
-                    )
-
-                    expect(fraudChallenge.closed).to.equal(true)
-                  })
-
-                  it("should send the ether deposited by the challenger to the treasury", async () => {
-                    await expect(tx).to.changeEtherBalance(
-                      bridge,
-                      (await bridge.fraudChallengeDepositAmount()).mul(-1)
-                    )
-                    await expect(tx).to.changeEtherBalance(
-                      treasury,
-                      await bridge.fraudChallengeDepositAmount()
-                    )
-                  })
-
-                  it("should emit FraudChallengeDefeated event", async () => {
-                    await expect(tx)
-                      .to.emit(bridge, "FraudChallengeDefeated")
-                      .withArgs(
-                        fraudWalletPublicKeyHash,
-                        data.sighash,
-                        data.signature.v,
-                        data.signature.r,
-                        data.signature.s
-                      )
-                  })
-                }
-              )
-
-              context(
-                "when the inputs are not marked as correctly spent in the Bridge",
-                () => {
-                  const data = nonWitnessSignMultipleInputsTx
-
-                  before(async () => {
-                    await createSnapshot()
-
-                    await bridge.setWallet(fraudWalletPublicKeyHash, {
-                      ecdsaWalletID: ethers.constants.HashZero,
-                      mainUtxoHash: ethers.constants.HashZero,
-                      pendingRedemptionsValue: 0,
-                      createdAt: await lastBlockTime(),
-                      state: walletState.Live,
-                    })
-
-                    await bridge
-                      .connect(thirdParty)
-                      .submitFraudChallenge(
-                        fraudWalletPublicKey,
-                        data.sighash,
-                        data.signature,
-                        {
-                          value: await bridge.fraudChallengeDepositAmount(),
-                        }
-                      )
-                  })
-
-                  after(async () => {
-                    await restoreSnapshot()
-                  })
-
-                  it("should revert", async () => {
-                    await expect(
-                      bridge
-                        .connect(thirdParty)
-                        .defeatFraudChallenge(
-                          fraudWalletPublicKey,
-                          data.preimage,
-                          data.signature,
-                          false
-                        )
-                    ).to.be.revertedWith(
-                      "Spent UTXO not found among correctly spent UTXOs"
-                    )
-                  })
-                }
-              )
-            })
-          })
-
-          context("when the input is witness", () => {
-            context("when the transaction has single input", () => {
-              context(
-                "when the inputs are marked correctly spent in the Bridge",
-                () => {
-                  const data = witnessSignSingleInputTx
-                  let tx: Transaction
-
-                  before(async () => {
-                    await createSnapshot()
-
-                    await bridge.setWallet(fraudWalletPublicKeyHash, {
-                      ecdsaWalletID: ethers.constants.HashZero,
-                      mainUtxoHash: ethers.constants.HashZero,
-                      pendingRedemptionsValue: 0,
-                      createdAt: await lastBlockTime(),
-                      state: walletState.Live,
-                    })
-                    await bridge.setSweptDeposits(data.deposits)
-                    await bridge.setSpentMainUtxos(data.spentMainUtxos)
-
-                    await bridge
-                      .connect(thirdParty)
-                      .submitFraudChallenge(
-                        fraudWalletPublicKey,
-                        data.sighash,
-                        data.signature,
-                        {
-                          value: await bridge.fraudChallengeDepositAmount(),
-                        }
-                      )
-
-                    tx = await bridge
-                      .connect(thirdParty)
-                      .defeatFraudChallenge(
-                        fraudWalletPublicKey,
-                        data.preimage,
-                        data.signature,
-                        true
-                      )
-                  })
-
-                  after(async () => {
-                    await restoreSnapshot()
-                  })
-
-                  it("should mark the challenge as closed", async () => {
-                    const challengeKey = buildChallengeKey(
-                      fraudWalletPublicKey,
-                      data.sighash,
-                      data.signature.v,
-                      data.signature.r,
-                      data.signature.s
-                    )
-
-                    const fraudChallenge = await bridge.fraudChallenges(
-                      challengeKey
-                    )
-
-                    expect(fraudChallenge.closed).to.equal(true)
-                  })
-
-                  it("should send the ether deposited by the challenger to the treasury", async () => {
-                    await expect(tx).to.changeEtherBalance(
-                      bridge,
-                      (await bridge.fraudChallengeDepositAmount()).mul(-1)
-                    )
-                    await expect(tx).to.changeEtherBalance(
-                      treasury,
-                      await bridge.fraudChallengeDepositAmount()
-                    )
-                  })
-
-                  it("should emit FraudChallengeDefeated event", async () => {
-                    await expect(tx)
-                      .to.emit(bridge, "FraudChallengeDefeated")
-                      .withArgs(
-                        fraudWalletPublicKeyHash,
-                        data.sighash,
-                        data.signature.v,
-                        data.signature.r,
-                        data.signature.s
-                      )
-                  })
-                }
-              )
-
-              context(
-                "when the inputs are not marked as correctly spent in the Bridge",
-                () => {
-                  const data = witnessSignSingleInputTx
-
-                  before(async () => {
-                    await createSnapshot()
-
-                    await bridge.setWallet(fraudWalletPublicKeyHash, {
-                      ecdsaWalletID: ethers.constants.HashZero,
-                      mainUtxoHash: ethers.constants.HashZero,
-                      pendingRedemptionsValue: 0,
-                      createdAt: await lastBlockTime(),
-                      state: walletState.Live,
-                    })
-
-                    await bridge
-                      .connect(thirdParty)
-                      .submitFraudChallenge(
-                        fraudWalletPublicKey,
-                        data.sighash,
-                        data.signature,
-                        {
-                          value: await bridge.fraudChallengeDepositAmount(),
-                        }
-                      )
-                  })
-
-                  after(async () => {
-                    await restoreSnapshot()
-                  })
-
-                  it("should revert", async () => {
-                    await expect(
-                      bridge
-                        .connect(thirdParty)
-                        .defeatFraudChallenge(
-                          fraudWalletPublicKey,
-                          data.preimage,
-                          data.signature,
-                          true
-                        )
-                    ).to.be.revertedWith(
-                      "Spent UTXO not found among correctly spent UTXOs"
-                    )
-                  })
-                }
-              )
-            })
-
-            context("when the transaction has multiple inputs", () => {
-              context(
-                "when the inputs are marked correctly spent in the Bridge",
-                () => {
-                  const data = witnessSignMultipleInputTx
-                  let tx: Transaction
-
-                  before(async () => {
-                    await createSnapshot()
-
-                    await bridge.setWallet(fraudWalletPublicKeyHash, {
-                      ecdsaWalletID: ethers.constants.HashZero,
-                      mainUtxoHash: ethers.constants.HashZero,
-                      pendingRedemptionsValue: 0,
-                      createdAt: await lastBlockTime(),
-                      state: walletState.Live,
-                    })
-                    await bridge.setSweptDeposits(data.deposits)
-                    await bridge.setSpentMainUtxos(data.spentMainUtxos)
-
-                    await bridge
-                      .connect(thirdParty)
-                      .submitFraudChallenge(
-                        fraudWalletPublicKey,
-                        data.sighash,
-                        data.signature,
-                        {
-                          value: await bridge.fraudChallengeDepositAmount(),
-                        }
-                      )
-
-                    tx = await bridge
-                      .connect(thirdParty)
-                      .defeatFraudChallenge(
-                        fraudWalletPublicKey,
-                        data.preimage,
-                        data.signature,
-                        true
-                      )
-                  })
-
-                  after(async () => {
-                    await restoreSnapshot()
-                  })
-
-                  it("should mark the challenge as closed", async () => {
-                    const challengeKey = buildChallengeKey(
-                      fraudWalletPublicKey,
-                      data.sighash,
-                      data.signature.v,
-                      data.signature.r,
-                      data.signature.s
-                    )
-
-                    const fraudChallenge = await bridge.fraudChallenges(
-                      challengeKey
-                    )
-
-                    expect(fraudChallenge.closed).to.equal(true)
-                  })
-
-                  it("should send the ether deposited by the challenger to the treasury", async () => {
-                    await expect(tx).to.changeEtherBalance(
-                      bridge,
-                      (await bridge.fraudChallengeDepositAmount()).mul(-1)
-                    )
-                    await expect(tx).to.changeEtherBalance(
-                      treasury,
-                      await bridge.fraudChallengeDepositAmount()
-                    )
-                  })
-
-                  it("should emit FraudChallengeDefeated event", async () => {
-                    await expect(tx)
-                      .to.emit(bridge, "FraudChallengeDefeated")
-                      .withArgs(
-                        fraudWalletPublicKeyHash,
-                        data.sighash,
-                        data.signature.v,
-                        data.signature.r,
-                        data.signature.s
-                      )
-                  })
-                }
-              )
-
-              context(
-                "when the inputs are not marked as correctly spent in the Bridge",
-                () => {
-                  const data = witnessSignMultipleInputTx
-
-                  before(async () => {
-                    await createSnapshot()
-
-                    await bridge.setWallet(fraudWalletPublicKeyHash, {
-                      ecdsaWalletID: ethers.constants.HashZero,
-                      mainUtxoHash: ethers.constants.HashZero,
-                      pendingRedemptionsValue: 0,
-                      createdAt: await lastBlockTime(),
-                      state: walletState.Live,
-                    })
-
-                    await bridge
-                      .connect(thirdParty)
-                      .submitFraudChallenge(
-                        fraudWalletPublicKey,
-                        data.sighash,
-                        data.signature,
-                        {
-                          value: await bridge.fraudChallengeDepositAmount(),
-                        }
-                      )
-                  })
-
-                  after(async () => {
-                    await restoreSnapshot()
-                  })
-
-                  it("should revert", async () => {
-                    await expect(
-                      bridge
-                        .connect(thirdParty)
-                        .defeatFraudChallenge(
-                          fraudWalletPublicKey,
-                          data.preimage,
-                          data.signature,
-                          true
-                        )
-                    ).to.be.revertedWith(
-                      "Spent UTXO not found among correctly spent UTXOs"
-                    )
-                  })
-                }
-              )
-            })
-          })
-        })
-
-        context("when the sighash type is incorrect", () => {
-          const data = wrongSighashType
-
-          before(async () => {
-            await createSnapshot()
-
-            await bridge.setWallet(fraudWalletPublicKeyHash, {
-              ecdsaWalletID: ethers.constants.HashZero,
-              mainUtxoHash: ethers.constants.HashZero,
-              pendingRedemptionsValue: 0,
-              createdAt: await lastBlockTime(),
-              state: walletState.Live,
-            })
-            await bridge.setSweptDeposits(data.deposits)
-            await bridge.setSpentMainUtxos(data.spentMainUtxos)
-
-            await bridge
-              .connect(thirdParty)
-              .submitFraudChallenge(
-                fraudWalletPublicKey,
-                data.sighash,
-                data.signature,
-                {
-                  value: await bridge.fraudChallengeDepositAmount(),
-                }
-              )
-          })
-
-          after(async () => {
-            await restoreSnapshot()
-          })
-
-          it("should revert", async () => {
-            await expect(
-              bridge
-                .connect(thirdParty)
-                .defeatFraudChallenge(
-                  fraudWalletPublicKey,
-                  data.preimage,
-                  data.signature,
-                  true
-                )
-            ).to.be.revertedWith("Wrong sighash type")
-          })
-        })
-      })
-
-      context("when the challenge is closed by defeat", () => {
-        const data = nonWitnessSignSingleInputTx
-
-        before(async () => {
-          await createSnapshot()
-
-          await bridge.setWallet(fraudWalletPublicKeyHash, {
-            ecdsaWalletID: ethers.constants.HashZero,
-            mainUtxoHash: ethers.constants.HashZero,
-            pendingRedemptionsValue: 0,
-            createdAt: await lastBlockTime(),
-            state: walletState.Live,
-          })
-          await bridge.setSweptDeposits(data.deposits)
-          await bridge.setSpentMainUtxos(data.spentMainUtxos)
-
-          await bridge
-            .connect(thirdParty)
-            .submitFraudChallenge(
-              fraudWalletPublicKey,
-              data.sighash,
-              data.signature,
-              {
-                value: await bridge.fraudChallengeDepositAmount(),
-              }
-            )
-
-          await bridge
-            .connect(thirdParty)
-            .defeatFraudChallenge(
-              fraudWalletPublicKey,
-              data.preimage,
-              data.signature,
-              false
-            )
-        })
-
-        after(async () => {
-          await restoreSnapshot()
-        })
-
-        it("should revert", async () => {
-          await expect(
-            bridge
-              .connect(thirdParty)
-              .defeatFraudChallenge(
-                fraudWalletPublicKey,
-                data.preimage,
-                data.signature,
-                false
-              )
-          ).to.be.revertedWith("Fraud challenge closed")
-        })
-      })
-
-      context("when the challenge is closed by timeout", () => {
-        const data = nonWitnessSignSingleInputTx
-
-        before(async () => {
-          await createSnapshot()
-
-          await bridge.setWallet(fraudWalletPublicKeyHash, {
-            ecdsaWalletID: ethers.constants.HashZero,
-            mainUtxoHash: ethers.constants.HashZero,
-            pendingRedemptionsValue: 0,
-            createdAt: await lastBlockTime(),
-            state: walletState.Live,
-          })
-          await bridge.setSweptDeposits(data.deposits)
-          await bridge.setSpentMainUtxos(data.spentMainUtxos)
-
-          await bridge
-            .connect(thirdParty)
-            .submitFraudChallenge(
-              fraudWalletPublicKey,
-              data.sighash,
-              data.signature,
-              {
-                value: await bridge.fraudChallengeDepositAmount(),
-              }
-            )
-
-          await increaseTime(await bridge.fraudChallengeDefeatTimeout())
-
-          await bridge
-            .connect(thirdParty)
-            .notifyFraudChallengeDefeatTimeout(
-              fraudWalletPublicKey,
-              data.sighash,
-              data.signature
-            )
-        })
-
-        after(async () => {
-          await restoreSnapshot()
-        })
-
-        it("should revert", async () => {
-          await expect(
-            bridge
-              .connect(thirdParty)
-              .defeatFraudChallenge(
-                fraudWalletPublicKey,
-                data.preimage,
-                data.signature,
-                false
-              )
-          ).to.be.revertedWith("Fraud challenge closed")
-        })
-      })
-    })
-
-    context("when the challenge does not exist", () => {
-      const data = nonWitnessSignMultipleInputsTx
-
-      before(async () => {
-        await createSnapshot()
-      })
-
-      after(async () => {
-        await restoreSnapshot()
-      })
-
-      it("should revert", async () => {
-        await expect(
-          bridge
-            .connect(thirdParty)
-            .defeatFraudChallenge(
-              fraudWalletPublicKey,
-              data.preimage,
-              data.signature,
-              false
-            )
-        ).to.be.revertedWith("Fraud challenge does not exist")
-      })
-    })
-  })
-
-  describe("notifyFraudChallengeDefeatTimeout", () => {
-    const data = nonWitnessSignSingleInputTx
-
-    describe("when the fraud challenge exists", () => {
-      describe("when the fraud challenge is open", () => {
-        describe("when the fraud challenge has timed out", () => {
-          let tx: Transaction
-
-          before(async () => {
-            await createSnapshot()
-
-            await bridge.setWallet(fraudWalletPublicKeyHash, {
-              ecdsaWalletID: ethers.constants.HashZero,
-              mainUtxoHash: ethers.constants.HashZero,
-              pendingRedemptionsValue: 0,
-              createdAt: await lastBlockTime(),
-              state: walletState.Live,
-            })
-            await bridge.setSweptDeposits(data.deposits)
-            await bridge.setSpentMainUtxos(data.spentMainUtxos)
-
-            await bridge
-              .connect(thirdParty)
-              .submitFraudChallenge(
-                fraudWalletPublicKey,
-                data.sighash,
-                data.signature,
-                {
-                  value: await bridge.fraudChallengeDepositAmount(),
-                }
-              )
-
-            await increaseTime(await bridge.fraudChallengeDefeatTimeout())
-
-            tx = await bridge
-              .connect(thirdParty)
-              .notifyFraudChallengeDefeatTimeout(
-                fraudWalletPublicKey,
-                data.sighash,
-                data.signature
-              )
-          })
-
-          after(async () => {
-            await restoreSnapshot()
-          })
-
-          it("should mark the fraud challenge as closed", async () => {
-            const challengeKey = buildChallengeKey(
-              fraudWalletPublicKey,
-              data.sighash,
-              data.signature.v,
-              data.signature.r,
-              data.signature.s
-            )
-
-            const fraudChallenge = await bridge.fraudChallenges(challengeKey)
-
-            expect(fraudChallenge.closed).to.be.true
-          })
-
-          it("should return the deposited ether to the challenger", async () => {
-            await expect(tx).to.changeEtherBalance(
-              bridge,
-              (await bridge.fraudChallengeDepositAmount()).mul(-1)
-            )
-            await expect(tx).to.changeEtherBalance(
-              thirdParty,
-              await bridge.fraudChallengeDepositAmount()
-            )
-          })
-
-          it("should emit FraudChallengeDefeatTimeout event", async () => {
-            await expect(tx)
-              .to.emit(bridge, "FraudChallengeDefeatTimeout")
-              .withArgs(
-                fraudWalletPublicKeyHash,
-                data.sighash,
-                data.signature.v,
-                data.signature.r,
-                data.signature.s
-              )
-          })
-        })
-
-        describe("when the fraud challenge has not timed out yet", () => {
-          before(async () => {
-            await createSnapshot()
-
-            await bridge.setWallet(fraudWalletPublicKeyHash, {
-              ecdsaWalletID: ethers.constants.HashZero,
-              mainUtxoHash: ethers.constants.HashZero,
-              pendingRedemptionsValue: 0,
-              createdAt: await lastBlockTime(),
-              state: walletState.Live,
-            })
-            await bridge.setSweptDeposits(data.deposits)
-            await bridge.setSpentMainUtxos(data.spentMainUtxos)
-
-            await bridge
-              .connect(thirdParty)
-              .submitFraudChallenge(
-                fraudWalletPublicKey,
-                data.sighash,
-                data.signature,
-                {
-                  value: await bridge.fraudChallengeDepositAmount(),
-                }
-              )
-
-            await increaseTime(
-              (await bridge.fraudChallengeDefeatTimeout()).sub(2)
-            )
-          })
-
-          after(async () => {
-            await restoreSnapshot()
-          })
-
-          it("should revert", async () => {
-            await expect(
-              bridge
-                .connect(thirdParty)
-                .notifyFraudChallengeDefeatTimeout(
-                  fraudWalletPublicKey,
-                  data.sighash,
-                  data.signature
-                )
-            ).to.be.revertedWith(
-              "Fraud challenge defeat period did not time out yet"
-            )
-          })
-        })
-      })
-
-      describe("when the fraud challenge is closed by challenge defeat", () => {
-        before(async () => {
-          await createSnapshot()
-
-          await bridge.setWallet(fraudWalletPublicKeyHash, {
-            ecdsaWalletID: ethers.constants.HashZero,
-            mainUtxoHash: ethers.constants.HashZero,
-            pendingRedemptionsValue: 0,
-            createdAt: await lastBlockTime(),
-            state: walletState.Live,
-          })
-          await bridge.setSweptDeposits(data.deposits)
-          await bridge.setSpentMainUtxos(data.spentMainUtxos)
-
-          await bridge
-            .connect(thirdParty)
-            .submitFraudChallenge(
-              fraudWalletPublicKey,
-              data.sighash,
-              data.signature,
-              {
-                value: await bridge.fraudChallengeDepositAmount(),
-              }
-            )
-
-          await bridge
-            .connect(thirdParty)
-            .defeatFraudChallenge(
-              fraudWalletPublicKey,
-              data.preimage,
-              data.signature,
-              false
-            )
-        })
-
-        after(async () => {
-          await restoreSnapshot()
-        })
-
-        it("should revert", async () => {
-          await expect(
-            bridge
-              .connect(thirdParty)
-              .notifyFraudChallengeDefeatTimeout(
-                fraudWalletPublicKey,
-                data.sighash,
-                data.signature
-              )
-          ).to.be.revertedWith("Fraud challenge is closed")
-        })
-      })
-
-      describe("when the fraud challenge is closed by previous timeout notification", () => {
-        before(async () => {
-          await createSnapshot()
-
-          await bridge.setWallet(fraudWalletPublicKeyHash, {
-            ecdsaWalletID: ethers.constants.HashZero,
-            mainUtxoHash: ethers.constants.HashZero,
-            pendingRedemptionsValue: 0,
-            createdAt: await lastBlockTime(),
-            state: walletState.Live,
-          })
-          await bridge.setSweptDeposits(data.deposits)
-          await bridge.setSpentMainUtxos(data.spentMainUtxos)
-
-          await bridge
-            .connect(thirdParty)
-            .submitFraudChallenge(
-              fraudWalletPublicKey,
-              data.sighash,
-              data.signature,
-              {
-                value: await bridge.fraudChallengeDepositAmount(),
-              }
-            )
-
-          await increaseTime(await bridge.fraudChallengeDefeatTimeout())
-
-          await bridge
-            .connect(thirdParty)
-            .notifyFraudChallengeDefeatTimeout(
-              fraudWalletPublicKey,
-              data.sighash,
-              data.signature
-            )
-        })
-
-        after(async () => {
-          await restoreSnapshot()
-        })
-
-        it("should revert", async () => {
-          await expect(
-            bridge
-              .connect(thirdParty)
-              .notifyFraudChallengeDefeatTimeout(
-                fraudWalletPublicKey,
-                data.sighash,
-                data.signature
-              )
-          ).to.be.revertedWith("Fraud challenge is closed")
-        })
-      })
-    })
-
-    describe("when the fraud challenge does not exist", () => {
-      before(async () => {
-        await createSnapshot()
-      })
-
-      after(async () => {
-        await restoreSnapshot()
-      })
-
-      it("should revert", async () => {
-        await expect(
-          bridge
-            .connect(thirdParty)
-            .notifyFraudChallengeDefeatTimeout(
-              fraudWalletPublicKey,
-              data.sighash,
-              data.signature
-            )
-        ).to.be.revertedWith("Fraud challenge does not exist")
-      })
-    })
-  })
-
   async function runSweepScenario(
     data: SweepTestData
   ): Promise<ContractTransaction> {
@@ -6383,19 +4986,6 @@ describe("Bridge", () => {
     return ethers.utils.solidityKeccak256(
       ["bytes20", "bytes"],
       [walletPubKeyHash, redeemerOutputScript]
-    )
-  }
-
-  function buildChallengeKey(
-    walletPublicKey: BytesLike,
-    sighash: BytesLike,
-    v: number,
-    r: BytesLike,
-    s: BytesLike
-  ): string {
-    return ethers.utils.solidityKeccak256(
-      ["bytes", "bytes32", "uint8", "bytes32", "bytes32"],
-      [walletPublicKey, sighash, v, r, s]
     )
   }
